+++ modules/meteroid/crates/meteroid-store/src/repositories/api_tokens.rs
@@ -1,3 +1,9 @@
+use argon2::{
+    password_hash::{rand_core::OsRng, PasswordHasher, SaltString},
+    Argon2,
+};
+use nanoid::nanoid;
+use tracing_log::log;
 use uuid::Uuid;
 
 use crate::domain::api_tokens::ApiToken;
@@ -13,7 +19,7 @@ pub trait ApiTokensInterface {
     ) -> StoreResult<Vec<ApiToken>>;
     async fn get_api_token_by_id(&self, id: &uuid::Uuid) -> StoreResult<ApiToken>;
 
-    async fn insert_api_token(&self, plan: domain::ApiTokenNew) -> StoreResult<ApiToken>;
+    async fn insert_api_token(&self, plan: domain::ApiTokenNew) -> StoreResult<(String, ApiToken)>;
 }
 
 #[async_trait::async_trait]
@@ -42,15 +48,65 @@ impl ApiTokensInterface for Store {
         Ok(api_token.into())
     }
 
-    async fn insert_api_token(&self, entity: domain::ApiTokenNew) -> StoreResult<ApiToken> {
+    async fn insert_api_token(
+        &self,
+        entity: domain::ApiTokenNew,
+    ) -> StoreResult<(String, ApiToken)> {
         let mut conn = self.get_conn().await?;
 
-        let insertable_entity: diesel_models::api_tokens::ApiTokenNew = entity.into();
+        //let id = uuid_gen::v7();
+        let id = common_utils::uuid::v7();
 
-        insertable_entity
+        // TODO
+        // api key is ex: ${pv for private key ?? pb for publishable key}_${tenant.env}_ + random
+        let prefix = "pv_sand_";
+
+        // encode in base62. Identifier is added to the api key, and used to retrieve the hash.
+        let id_part = base62::encode(id.as_u128());
+
+        // Generate the api key
+        let api_key_random = nanoid!(28, &common_utils::rng::BASE62_ALPHABET);
+        let api_key = format!("{}{}/{}", &prefix, &api_key_random, &id_part);
+
+        // Generate the hash that we will store in db
+        let argon2 = Argon2::new(
+            argon2::Algorithm::Argon2id,
+            argon2::Version::V0x13,
+            argon2::Params::new(5 * 1024, 1, 1, None).unwrap(),
+        );
+        let salt = SaltString::generate(&mut OsRng);
+        let api_key_hash = argon2
+            .hash_password(&api_key_random.as_bytes(), &salt)
+            .map_err(|e| {
+                log::error!("Unable to hash api key: {}", e);
+                StoreError::InvalidArgument("unable to hash api key".to_string())
+            })?
+            .to_string();
+
+        // generate a hint that will also be stored
+        let hint = format!(
+            "{}{}...{}",
+            &prefix,
+            &api_key_random[..4],
+            &id_part[id_part.len() - 4..]
+        );
+
+        let insertable_entity = diesel_models::api_tokens::ApiTokenNew {
+            id,
+            name: entity.name,
+            created_at: chrono::Utc::now().naive_utc(),
+            created_by: entity.created_by,
+            tenant_id: entity.tenant_id,
+            hash: api_key_hash,
+            hint: hint,
+        };
+
+        let result = insertable_entity
             .insert(&mut conn)
             .await
             .map_err(Into::into)
-            .map(Into::into)
+            .map(Into::into);
+
+        result.map(|res| (api_key, res))
     }
 }
